+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - **Added**: Primary color changes based on `ELVIRA_FACULTY`
 - **Changed**: `Required Auth/Admin` to `Auth/Admin Guard`
 - **Changed**: `Login.tsx` to `Auth.tsx`
-- **Changed**: `README.md`
+- **Changed**: `README.md` and cookies info
 - **Changed**: Viewer from `0.5.9` to `0.5.13`
 - **Changed**: In .env `ELVIRA_FACULTY` variable to `ELVIRA_THEME`
 - **Removed**: useCustomEffect